LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/zero-intrinsic-iteration-duration.tentative-expected.txt


FAIL Intrinsic iteration duration is non-negative assert_approx_equals: values do not match for "Duration is zero when start > end" expected 0 +/- 0.125 but got -33.333335876464844
